prescribing-info-app/pyproject.toml
[tool.poetry]
name = "writer-framework-default"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
writer = "^0.8.2"
bio = "^1.7.1"
bs4 = "^0.0.2"
python-dotenv = "^1.0.1"
asyncio = "^3.4.3"
aiohttp = "^3.11.11"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
